apps/terminal_api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pyodbc>=5.1.0
PyJWT>=2.10.0
passlib[bcrypt]>=1.7.4
